+++ serviceTest/CHANGELOG.md
@@ -3,9 +3,10 @@
 This file documents changes to the `workbench-service-test` library, including notes on how to upgrade to new versions.
 
 ## 2.0
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "2.0-1174fb6"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "2.0-TRAVIS-REPLACE-ME"`
 
 ### Changed
+- add ability to create rawls billing projects using Azure managed app coordinates
 - ensure http response entities are read only once
 - add optional param for ignoreEmptyColumns in rawls submission API
 - updated `rawls-model` dependency to `0.1-04a7a76b`